Add files via upload
Question 1 
PayMoney. processes thousands of transactions daily amounting to crores of Rupees. They
also have a daily target that they must achieve. Given a list of transactions done by
PayMoney and a daily target, your task is to determine at which transaction PayMoney
achieves the same. If the target is not achievable, then display the target is not achieved.